:sparkles: (ECS) improve typing of components and entities

+++ lib/ecs/components/movable.dart
@@ -8,8 +8,8 @@ class Movable extends Component {
 }
 
 class Speed {
-  double dx;
-  double dy;
+  int dx;
+  int dy;
 
   Speed({this.dx = 0, this.dy = 0})
       : assert(dx != null),
